* Moved destroyBlock and playBlockDestructionEffect to ApiBlock
